update conky2.conkyrc2
The  file that I uploaded was edited to bring it up to the current syntax. Unfortunuately, after downloading and testing it again, it does not work at all. This is the unedited file in the outdated syntax and seems to work fine. I cannot explain it but THIS is the file that should properly load with the clock and date.